% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allpointplot.R
\name{allpointplot}
\alias{allpointplot}
\title{plots simulation data with samplle size on x-axis}
\usage{
allpointplot(data)
}
\arguments{
\item{data}{list; first simulation data as returned by changingn.(...) or changingvar.(...) functions.}
}
\value{
a data.frame representing the input data, where the rows are the different measures and the columns are the datapoints given by the simulation data (for example stepsizes or different standard deviations).
}
\description{
plots simulation data with sample size on x-axis and every datapoint in the plot
}
\details{
#Takes Simulation data, as produced by for example changingn.daypattern as input, those need to have a roq named n where the sample sizes for the calculation are stored in. This function will plot the results of the measures on the y axis and the sample sizes on the x axis.
}
\examples{

bla <-changingn.daypattern(stepsizes = c(2,-4), repeatings = 2)
allpointplot(bla)

}
\seealso{
\code{scale}
}
\author{
Florian Berger <florian_berger@ymail.com>
}
