% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetaData.R
\name{BetaData}
\alias{BetaData}
\title{Creates abundance dataset from beta distribution}
\usage{
BetaData(n, shape1, shape2, StartAndEnd, FillUp = TRUE)
}
\arguments{
\item{n}{numeric, whole number; number of drawings to be done from the normal distribution, also the number of observations in the whole dataset.}

\item{shape1}{numeric; Parameter for underlying beta-distribution.}

\item{shape2}{numeric; Parameter for underlying beta-distribution.}

\item{StartAndEnd}{vector; Start and end values in which the dates will be. If those are not date objects, they will be interpreted as difference from the 01-01-1970.}

\item{FillUp}{logical; If TRUE not specified datapoints will be filled with 0.}
}
\value{
a dataframe with two columns; one with date objects, one with the number of observations on the respective dates.
}
\description{
Creates abundance dataset from beta distribution with given parameters.
}
\details{
Creates a data.frame, with two columns, one called dates and one called count. It will get n realizations of a beta distributions with the parameters given by shape 1 and shape 2.Those realizations will be converted, that the range of the realizations is no longer between zero and one, but the range between two dates that should be given by StartAndEnd. It will round those to whole numbers and convert them into dates, by treating these realizations as differences from the 2017-01-01. If fillup = TRUE it will fillup the dates that are in between the minimum and maximum of the dates, and give count values of zero for those days.
}
\examples{

dat  <- BetaData( shape1 = 5,shape2 = 1, n = 1000, StartAndEnd = c(0,15))
func <- as.function(estdist(dat))
curve(func, from = as.numeric(min(dat$Date)), to = as.numeric(max(dat$Date)))

}
\seealso{
\code{scale}
}
\author{
Florian Berger <florian_berger@ymail.com>
}
