% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changingn-normaldist.R
\name{changingn.normaldist}
\alias{changingn.normaldist}
\title{simulates abundance data and calculates overlap for changing sample sizes}
\usage{
changingn.normaldist(ns = c(1000, 500, 200, 100, 50, 20, 10),
  repeatings = 100, type = "beta", OLest = c("Weitzman", "Matusita",
  "Pianka", "Morisita", "Duration", "WMD", "Horn", "n"))
}
\arguments{
\item{ns}{vector of numeric objects (whole numbers); the sample sizes for which the overlap will be determined. Applies for both datasets.}

\item{repeatings}{numeric (whole numbers); Determines how often the whole procedure (data producing and calculation of the overlap) shall be repeated.}

\item{type}{string; If set to beta the abundance data will bbe drawn from a beta distribution. If not beta this will be from a normal distribution.}

\item{OLest}{vector of strings; determines in which way(s) the overlap shall be determined, will be passed to OLE function.}
}
\value{
a list. First index level determines the way of estimating the pdf so it has the three indexes: 'kernel', 'fitdistr' and 'normalize'. The second index-level is the sample size and the last the number of the repeating. In there are the results as returned by OLE function. To illustrate: list[[kernel]][[sample-size]][[repeating]]
}
\description{
simulates abundance data and calculates overlap for changing sample sizes
}
\details{
Produces two abundance datastes and dteremines the overlpa-measures given with OLest betweeen these. The overlap will be determined using three different methods : kernel density estimation, fitting a distribution or normalization (normalize function). The sample size will be set to the maximum value of ns. The type argument determinnes whether these data shall be created using a beta or a normal distribution. Then the sample size will be decreased, using the normaldist.crossout function, to  the next lower value of ns, and the overlap will be determined again. This will be repeated, until an overlap for all the ns was calculated. Afterwards This procedure itself will be repeated as many times as repeatings says. A listt will be returned, that has 3 lists in it, one for the kernel ddensity estimation one for fitted distributions etc. This lists have a list stored in them for each stepsize. Those lists contain all the results of the estimations for the respective sample size and way the overlap was determined.
}
\examples{
changingn.normaldist(ns=c(300,50), repeatings = 2, type = 'normal')

}
\seealso{
\code{scale}
}
\author{
Florian Berger <florian_berger@ymail.com>
}
