% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changingvar-beta.R
\name{changingvar.beta}
\alias{changingvar.beta}
\title{simulates abundance data and calculates overlap for changing variances}
\usage{
changingvar.beta(repeatings = 100, days = c(730, 680, 550, 365, 180, 90, 7),
  n = 1000)
}
\arguments{
\item{repeatings}{numeric (whole numbers); Determines how often the whole procedure (data producing and calculation of the overlap) shall be repeated.}

\item{days}{vector of numerics (whole numbers); determines the length of the time range over which the data shall stretch.}

\item{n}{numeric (whole number); Determines the sample size for creating the datasets. Will be passed as as 'n' argument to to the NormData() function.}
}
\value{
a list. First index level determines the way of estimating the pdf so it has the three indexes: 'kernel', 'fitdistr' and 'normalize'. The second index-level is the chosen numbers of days and the last the number of the repeating. In there are the results as returned by OLE function. To illustrate: list[[kernel]][[numbers of days]][[repeating]].
}
\description{
simulates abundance data and calculates overlap for changing variances
}
\details{
Will determine the overlap-measures given by OLest for two simulated datasets both coming from a beta distribution with the sample size given by n. This will be done for every number of days given by days which determines the time-range over which the dataset spreads. And for every number of day there will be as many overlaps will be determined as the repeatings argument says.
}
\examples{
changingvar.beta(days=c(5,12), n = 500, repeatings = 2)

}
\seealso{
\code{scale}
}
\author{
Florian Berger <florian_berger@ymail.com>
}
