% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckData.R
\name{CheckData}
\alias{CheckData}
\title{checks input data for pdf-estimation and some overlap-estimators}
\usage{
CheckData(data, normalize = FALSE, Interpolate = TRUE)
}
\arguments{
\item{data}{data-frame; Abundance dataset. It must have two columns, one wih date-objects one with numeric values.}

\item{normalize}{logical; determines, whether data will be normalized or if probabilities will be estimated in another way. Runs additional checks if TRUE.}

\item{interpolate}{logical; determines whether data should be interpolated later on or not. Will run additional Checks if TRUE.}
}
\value{
data.frame. Same as data(input) but with renamed columns, one named Date (the one with date objects) and one named Count.
}
\description{
checks input data for pdf-estimation and some overlap-estimators
}
\details{
This function Checks if the data fits the requirements for the distribution estimation and for some of the Overlapmeasures As it Checks the data, the data argument can be everything, but it will only be returned again, if it's a dataframe with two columns, one with dates and one with numerics It has an Dataframe as outcome, which is exactly like the data that is given but the two columns are called Date and Count
}
\examples{

a <- NormData(mean=0, sd =20 ,n=1000, FillUp = FALSE)
newa <- FillUp(a, NA)

}
\seealso{
\code{scale}
}
\author{
Florian Berger <florian_berger@ymail.com>
}
