% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckHurlbert.R
\name{CheckHurlbert}
\alias{CheckHurlbert}
\title{checks input data for OLE.hurlbert()}
\usage{
CheckHurlbert(data, Interpolate = TRUE)
}
\arguments{
\item{data}{data-frame; It must have two columns, one wih date-objects one with numeric values.}

\item{interpolate}{logical; determines whether data should be interpolated or not.}
}
\value{
data.frame. based on input data.frame, may include new datapoints.
}
\description{
checks input data for OLE.hurlbert()
}
\details{
This function checks if given data fits the requirenments for the OLE.hurlbert() function. Will interpolate the data if required.
}
\examples{

a      <- c(1:10)
dates  <- as.Date(a, origin = '2017-01-01')
count  <- c(1,1,3,4,6,9,5,4,2,1)
dat    <- data.frame(dates, count)
newdat <- CheckHurlbert(dat)


}
\seealso{
\code{scale}
}
\author{
Florian Berger <florian_berger@ymail.com>
}
