% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckProbFunc.R
\name{checkProbFunc}
\alias{checkProbFunc}
\title{checks input data for EstimatorData() Function}
\usage{
checkProbFunc(data)
}
\arguments{
\item{data}{list; list that is to be checked, as returned by estdist() or uniform() function.}
}
\value{
nothing.
}
\description{
checks input data if it fits the requirenments of the EstimatorData() Function
}
\details{
This function Checks whether given data fits requirenments for the EstimatorData() function those need to have a $startpoint, $endpoint and $func,  $func needs to be a function It doesn't return anything. Use prior to EstimatorData() Function.
}
\examples{

a     <- c(1:10)
dates <- as.Date(a, origin = '2017-01-01')
count <- c(1,1,3,4,6,9,5,4,2,1)
dat   <- data.frame(dates, count)
checkProbFunc(estdist(dat, returnonlyfunction = FALSE))


}
\seealso{
\code{scale}
}
\author{
Florian Berger <florian_berger@ymail.com>
}
