% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountsToDrawing.R
\name{CountsToDrawing}
\alias{CountsToDrawing}
\title{Makes a vector of a abundance dataframe}
\usage{
CountsToDrawing(data, CheckData = TRUE)
}
\arguments{
\item{data}{data.frame;  abundance data, with one column of date objects and one with numeric values.}

\item{CheckData}{logical; If TRUE the function will check the data input whether it fits the requirements using the CheckData function. If FALSE it will process without, which might result in wrong results so to use the default value (TRUE) is strongly recommended.}
}
\value{
A vector where each date in the input data.frame is stored in as many times as there are counts in the same row in the numeric column.
}
\description{
Makes a vector with each date stored in it as many times as there were observations.
}
\details{
Makes a vector with each date stored in it as many times as there were observations.
}
\examples{
a     <- c(1:10)
dates <- as.Date(a, origin = '2017-01-01')
count <- c(1,1,3,4,6,9,5,4,2,1)
dat   <- data.frame(dates, count)
CountsToDrawing(dat)

}
\seealso{
\code{scale}
}
\author{
Florian Berger <florian_berger@ymail.com>
}
