% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daypattern-crossout.R
\name{daypattern.crossout}
\alias{daypattern.crossout}
\title{Crosses out days in a pattern from abundance dataset(s)}
\usage{
daypattern.crossout(data1, data2 = NULL, stepsize, CheckData = TRUE)
}
\arguments{
\item{data1}{data.frame; an abundance dataset, with one column of date objects, and one column numeric values.}

\item{data2}{data.frame; an abundance dataset, with one column of date objects, and one column numeric values. Can also be NULL.}

\item{stepsize}{numeric (whole numbers); determines the pattern in which the dates shall be crossed out. If stepsize positive it will exclude every Xth day, if negative every Xth day will be included.}

\item{CheckData}{logical; If TRUE the CheckData() function will be runned in advance. If FALSE the columns of data1 (and possibly data2) must be named Date and Count.}
}
\value{
If dataset2 i NULL a data.frame like data1, but with excluded days. If data2 i a data.frame a list with two levels: 1 is data1 2 is data2 both with th same days (same dates) excluded.
}
\description{
Crosses out days in a pattern from one or two abundance dataset
}
\details{
This method will cross out whole days in a pattern, meaning, that if stepsize is positive, every xth day will be crossed out, when its negative every xth day will be left in the dataset. If there is no second dataset given it will cross out the days for only one and return only one dataset otherwise it will cross out the exact same days. The startday for the pattern will be chosen randomly  in the interval of the first x days, where x is the value the stepsize argument is given.
}
\examples{

#Create Datasets
 a     <- c(1:10)
 dates1 <- as.Date(a, origin = '2017-01-01')
 count1 <- c(1,1,3,4,6,9,5,4,2,1)
 dat1   <- data.frame(dates1, count1)
 b     <- c(6:15)
 dates2 <- as.Date(b, origin = '2017-01-01')
 count2 <- c(1,2,4,5,6,9,3,4,1,1)
 dat2   <- data.frame(dates2, count2)
 
 daypattern.crossout(dat1, data2 = dat2, stepsize = -2, CheckData = TRUE)

}
\seealso{
\code{scale}
}
\author{
Florian Berger <florian_berger@ymail.com>
}
