% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FillUp.R
\name{FillUp}
\alias{FillUp}
\title{Fills up abundance data-set}
\usage{
FillUp(Data, Fillup = 0)
}
\arguments{
\item{data}{data-frame; Abundance dataset. It must have two columns, one wih date-objects one with numeric values.}

\item{FillUp}{numeric; This will be what will be filled up. for example 0, NA etc.}
}
\value{
data.frame. Same as data(input) but with filled up not specified dates.
}
\description{
Fills up not specified data-points (whole days) in an abundance dataset.
}
\details{
will fill up a abundance dataset, meaning that it will add the value given by the Fillup argument for all the dates, that are not existing in the dataset given by the Data argument. Only dates that are in between the minimum and the maximum of the dates in Data. the input data needs to have two columns, one named dates which should be filled with dates and count.
}
\examples{

a <- NormData(mean=0, sd =20 ,n=1000, FillUp = FALSE)
newa <- FillUp(a, NA)

}
\seealso{
\code{scale}
}
\author{
Florian Berger <florian_berger@ymail.com>
}
