% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcfromdata.R
\name{funcfromdata}
\alias{funcfromdata}
\title{basis to make a function from a dataset}
\usage{
funcfromdata(dat, xs, ys)
}
\arguments{
\item{dat}{numeric; X value for which a y-value shall be calculated by interpolating the data.}

\item{xs}{vector; vector of numeric values, 'x'component of dataset. In combination with ys this makes a dataset.}

\item{ys}{vector; vector of numeric values, 'y'component of dataset. In combination with xs this makes a dataset.}
}
\value{
a numeric value, representing the new y-value to the x-value givn by dat.
}
\description{
basis to make a function from a dataset
}
\details{
basis for creating a function from data, xs and ys is the basis for the calculation, so the function will return the interpolated y value to the x-value that is given by dat. the function will return 0 for Every value for dat, that is not in between the minimum and the maximum of xs.
}
\examples{

dat  <- BetaData( shape1 = 5,shape2 = 1, n = 1000, StartAndEnd = c(0,15))
func <- function(x)funcfromdata(x, xs = as.numeric(dat$Date), y = dat$Count)
curve(func, from = as.numeric(min(dat$Date)), to = as.numeric(max(dat$Date)))
 
}
\seealso{
\code{scale}
}
\author{
Florian Berger <florian_berger@ymail.com>
}
