% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdataframe.R
\name{getdataframe}
\alias{getdataframe}
\title{converts simulation data to a data.frame}
\usage{
getdataframe(data, fitdistr, sd = T)
}
\arguments{
\item{data}{list; first sublevel of simulation data as returned by changingn.(...) or changingvar.(...) functions. e.g. simulationdata[['kernel']]}

\item{fitdistr}{string; determines which data shall be plotted. Determines the first level of the input list and so it can either be 'kernel', 'distrest' or 'normalize'.}

\item{sd}{logical; If TRUE the function will give two columns for each datapoint, one with the means and one with the standard deviations of these means.}
}
\value{
a data.frame representing the input data, where the rows are the different measures and the columns are the datapoints given by the simulation data (for example stepsizes or different standard deviations).
}
\description{
converts simulation data to a data.frame
}
\details{
This function needs simulation data as produced by changingn.daypattern as input and gives a dataframe of the results for the probability density function estimation method (namely 'normalize', 'kernel' and 'fitdistr'). If sd = TRUE it will give two columns for each datapoint, one with the means and one with the standard deviations of these means.
}
\examples{

bla <-changingvar.beta(days=c(5,12), n = 500, repeatings = 2)

}
\seealso{
\code{scale}
}
\author{
Florian Berger <florian_berger@ymail.com>
}
