% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmean.R
\name{getmean}
\alias{getmean}
\title{gets mean for overlap measure for simulation dataset}
\usage{
getmean(data)
}
\arguments{
\item{data}{list; first sublevel of simulation data as returned by changingn.(...) or changingvar.(...) functions. e.g. simulationdata[['kernel']]}
}
\value{
a vector of numerrics, representing the means for each ddatapoint in the simulation data (e.g. smaple sizes) and has those as names.
}
\description{
gets mean for overlap measure for simulation dataset
}
\details{
needs simulation data, as produced by changingn.daypattern but only for one estimation method (kernel/normalize/fitdistr) and calculates the mean for every datapoint considering the measures measures : Weitzman\'s delta','Matusita\'s rho', 'Pianka\'s alpha','Morisitas\'s lambda',  'Horn\'s RO' returns a vector
}
\examples{

bla <-changingvar.beta(days=c(5,12), n = 500, repeatings = 2)
getmean(bla[['kernel']])

}
\seealso{
\code{scale}
}
\author{
Florian Berger <florian_berger@ymail.com>
}
