% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOLdata.R
\name{getOLdata}
\alias{getOLdata}
\title{extracts data for a single overlap-measure from simulation dataset}
\usage{
getOLdata(data, olname)
}
\arguments{
\item{data}{list; as returned by changingn.(...) or changingvar.(...) functions but with already determined pdf-estimation method, so e.g. simulationdata[['kernel']].}

\item{olname}{string; name of overlap-measure, as returned by OLE function. Must be in the data.}
}
\value{
a list as the input list but with the first two sublevels excluded.
}
\description{
extracts data for a single overlap-measure from simulation dataset
}
\details{
needs the simulation data, as produced by changingn.(...) or changingvar.(...) functions, for one estimation method('kernel/fitdistr/normalize) as input. Will return a list of all the results for the Overlpa-measure given by olname. This list will be seperated by the units the simulation function has (e.g. stepsizes, number of days. etc.)
}
\seealso{
\code{scale}
}
\author{
Florian Berger <florian_berger@ymail.com>
}
