% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelest.R
\name{kernelest}
\alias{kernelest}
\title{Runs Kernel Density Estimation for abundance data}
\usage{
kernelest(data, kernel = "gaussian", bandwidth = NULL,
  numberofdatapoints = 1024, CheckData = TRUE)
}
\arguments{
\item{kernel}{string; the kerel which shall be used. Can be "gaussian", "epanechnikov", "rectangular", "triangular" or "biweight".}

\item{bandwidth}{numeric; the bandwidth for the kernel density estimation. If is null it will be determined by the unbiased cross validation using the h.ucv() function from the 'kedd' package.}

\item{numberofdatapoints}{numeric, whole number; determines, how much datapoints shall be produced by the density() function. As can be seen in the help of density() this shall be a power of two.}

\item{CheckData}{logical; If TRUE the function will check the data input whether it fits the requirements using the CheckData function. If FALSE it will process without, which might result in wrong results so to use the default value (TRUE) is strongly recommended.}
}
\value{
returns a lit with 5 levels: data = results of the kernel density estimation (dataframe) with sublevels 'x' and 'y',estimationmethod = 'kernel density estimation', kernel = used kernel, bandwidth = used bandwidth, startpoint = first date in the data level, endpoint =  last point in data level)
}
\description{
This function will perform a Kernel Density Estimation
}
\details{
This function performs a simple Kernel Density estimation using the density() function. By default the bandwidth will be selected using the h.ucv() function from the 'kedd' package.
}
\examples{
a     <- c(1:10)
dates <- as.Date(a, origin = '2017-01-01')
count <- c(1,1,3,4,6,9,5,4,2,1)
dat   <- data.frame(dates, count)
kernelest(dat)

}
\seealso{
\code{scale}
}
\author{
Florian Berger <florian_berger@ymail.com>
}
