% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikelihood.R
\name{loglikelihood}
\alias{loglikelihood}
\title{Calculates the loglikelihood for an abundance datesset and a given function}
\usage{
loglikelihood(dates, func, allowDistrBorders = FALSE)
}
\arguments{
\item{dates}{vector;  dates, that should be in the range of the function.}

\item{func}{function object; a given function, for whih the loglikelihood shall be calculated.}

\item{allowDistrBorders}{logical; If TRUE the first and last values of the dates vector will b changed by 0.00001 (first) or -0.00001(last) if the corresponding result of the function is Inf or -Inf.}
}
\value{
A numeric value representing the loglikelihood.
}
\description{
Calculates the loglikelihood for an abundance datesset and a given function
}
\details{
Calculates the loglikelihood for an abundance datesset and a given function by multiplying all the values given by the given function, that uses the dates given by dates.
}
\examples{
a     <- c(1:10)
dates <- as.Date(a, origin = '2017-01-01')
count <- c(1,1,3,4,6,9,5,4,2,1)
dat   <- data.frame(dates, count)
loglikelihood(dates = dates, func = estdist(dat))

}
\seealso{
\code{scale}
}
\author{
Florian Berger <florian_berger@ymail.com>
}
