% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measureplot.R
\name{measureplot}
\alias{measureplot}
\title{saves plots from simulation data}
\usage{
measureplot(data, olmeasures)
}
\arguments{
\item{data}{list; first sublevel of simulation data as returned by changingn.(...) or changingvar.(...) functions. e.g. simulationdata[['kernel']]}

\item{olest}{string; name of overlap-measure, as returned by OLE function. Must be in the data.}

\item{logx}{logical; If TRUE x-axis will be logarithmic.}

\item{ns}{logical; If TRUE the x-axis will be the sample size and not the sublevel of the input data. Can only be TRUE if 'n' is in the data as overlap-measure.}

\item{diff}{logical; If TRUE it will plot the difference to the mean of the chosen datapoints, and not the absolute values.}

\item{file}{string; location where the files shall be saved. will be passed to jpeg() function.}
}
\value{
nothing but creates plot.
}
\description{
saves plots from simulation data
}
\details{
# takes simualtion data as produced by chaningn.daypattern as input. Thiss function will save all the plots using the OLplot function. It will save all the plots for the Overlap-Measures given by OLest. Will save files to directory given by file. if ns = TRUE it wwill use the n (if given as result in the data) on the x-axis instead of the daypattern or number of days. If diff= TRUE it will plot the distance to the mean of the respective datapoint. If logx = TRUE the x-axis will be logarithmic.
}
\examples{

#saves plots to current working directory

}
\seealso{
\code{scale}
}
\author{
Florian Berger <florian_berger@ymail.com>
}
