% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normaldist-crossout.R
\name{normaldist.crossout}
\alias{normaldist.crossout}
\title{Crosses out days from abundance dataset(s) using a normal distribution}
\usage{
normaldist.crossout(data, mean, sd, reduction = 1, DataCheck = TRUE)
}
\arguments{
\item{data}{data.frame; an abundance dataset, with one column of date objects, and one column numeric values.}

\item{mean}{numeric; the mean of the normal distribution. Number of days after the 01.01.1970}

\item{sd}{numeric; the standard deviation of the normal distribution.}

\item{reduction}{numeric; Determines how many 'observations' shall be excluded.}

\item{CheckData}{logical; If TRUE the CheckData() function will be runned in advance. If FALSE the columns of data1 (and possibly data2) must be named Date and Count.}
}
\value{
A data.frame like data, but with excluded Observations.
}
\description{
Crosses out days from abundance dataset(s) using a normal distribution
}
\details{
this function will reduce the number of observation in this dataset by one on a day, that is determined by a normal distribution, with the parameters mean and sd. if reduction is more than one, then there are several days determined.
}
\examples{

#Create Datasets
 a     <- c(1:10)
 dates1 <- as.Date(a, origin = '2017-01-01')
 count1 <- c(1,1,3,4,6,9,5,4,2,1)
 dat1   <- data.frame(dates1, count1)
 
 normaldist.crossout(dat1, mean = mean(as.numeric(dat1$dates1)), sd = 2, reduction =8)

}
\seealso{
\code{scale}
}
\author{
Florian Berger <florian_berger@ymail.com>
}
