% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OLE-duration.R
\name{OLE.duration}
\alias{OLE.duration}
\title{calculates the duration ratio}
\usage{
OLE.duration(data1, data2, CheckData = TRUE)
}
\arguments{
\item{data1}{data.frame or list; can either be an abundance dataset or an object for which the levels $startpoint and $endpoint are numeric values. The abundance dataset needs to have the levels $Date and $Count if CheckData = FALSE.}

\item{CheckData}{logical; If TRUE the CheckData() function will be runned in advance.}

\item{data1}{data.frame or list; can either be an abundance dataset or an object for which the levels $startpoint and $endpoint are numeric values. The abundance dataset needs to have the levels $Date and $Count if CheckData = FALSE.}
}
\value{
numeric value representing Weitzman's Delta.
}
\description{
calculates the duration ratio from two abundance-datasets
}
\details{
calculates the duration ratio, by dividing the duration that both datasets share by the duratio of both datasets (meaning from earliest point to the latest, if there i a continuity between the two datasets (if they share at least one datapoint)). The function takes abundance datasets as input. A list with the levels $startpoint and $endpoint can be used alternatively.
}
\examples{

#Create Datasets
 a     <- c(1:10)
 dates1 <- as.Date(a, origin = '2017-01-01')
 count1 <- c(1,1,3,4,6,9,5,4,2,1)
 dat1   <- data.frame(dates1, count1)
 b     <- c(6:15)
 dates2 <- as.Date(b, origin = '2017-01-01')
 count2 <- c(1,2,4,5,6,9,3,4,1,1)
 dat2   <- data.frame(dates2, count2)
 
 OLE.duration(dat1, dat2)

}
\seealso{
\code{scale}
}
\author{
Florian Berger <florian_berger@ymail.com>
}
