% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OLE-productmoment.R
\name{OLE.productmoment}
\alias{OLE.productmoment}
\title{calculates product-moment correlation coefficient}
\usage{
OLE.productmoment(data1, data2, CheckData = TRUE)
}
\arguments{
\item{data1}{data.frame; can either be an abundance dataset.}

\item{CheckData}{logical; if TRUE the CheckData() function will be run in advance.}

\item{data1}{data.frame; can either be an abundance dataset.}
}
\value{
numeric value representing product-moment correlation coefficient.
}
\description{
calculates product-moment correlation coefficient from two datasets
}
\details{
Calculates the product moment correlation coefficient as found in 'The Measurement of Niche Overlap and Some Relatives' (Hurlbert, S.H. 1971). The function takes abundance datasets as input.
}
\examples{

#Create Datasets
 a     <- c(1:10)
 dates1 <- as.Date(a, origin = '2017-01-01')
 count1 <- c(1,1,3,4,6,9,5,4,2,1)
 dat1   <- data.frame(dates1, count1)
 b     <- c(6:15)
 dates2 <- as.Date(b, origin = '2017-01-01')
 count2 <- c(1,2,4,5,6,9,3,4,1,1)
 dat2   <- data.frame(dates2, count2)

 OLE.productmoment(dat1,dat2)

}
\seealso{
\code{scale}
}
\author{
Florian Berger <florian_berger@ymail.com>
}
