% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OLE-wmd.R
\name{OLE.wmd}
\alias{OLE.wmd}
\title{calculates a 'weighted mean day' based overlap measure}
\usage{
OLE.wmd(data1, data2, CheckData = TRUE)
}
\arguments{
\item{data1}{data.frame; abundance dataset. Needs a column with only date objects and one with only numerics. If CheckData = TRUE those need to be named $Date and $Count.}

\item{data2}{data.frame; abundance dataset. Needs a column with only date objects and one with only numerics. If CheckData = TRUE those need to be named $Date and $Count. needs to have the levels $Date and $Count if CheckData = FALSE.}

\item{CheckData}{logical; If TRUE the CheckData() function will be runned in advance.}
}
\value{
numeric value representing the overlap.
}
\description{
calculates the overlap on the basis of the concept of the 'weighted mean day'
}
\details{
calculates the overlap on the basis of the concept of the 'weighted mean day' by dividing the difference of wmd by the square-root of the sum of the squared standard deviations of both datasets.
}
\examples{

#Create Datasets
 a     <- c(1:10)
 dates1 <- as.Date(a, origin = '2017-01-01')
 count1 <- c(1,1,3,4,6,9,5,4,2,1)
 dat1   <- data.frame(dates1, count1)
 b     <- c(6:15)
 dates2 <- as.Date(b, origin = '2017-01-01')
 count2 <- c(1,2,4,5,6,9,3,4,1,1)
 dat2   <- data.frame(dates2, count2)
 
 OLE.wmd(dat1, dat2)

}
\seealso{
\code{scale}
}
\author{
Florian Berger <florian_berger@ymail.com>
}
