% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageCheck.R
\name{PackageCheck}
\alias{PackageCheck}
\title{Checks if Package is already installed}
\usage{
PackageCheck(Package)
}
\arguments{
\item{Package}{string; Name of Package that is to be checked.}
}
\value{
attaches Package but returns nothing.
}
\description{
Checks if Package is already installed and attaches it if it is available.
}
\details{
This function checks whether given Package is installed, and if so it will use the library() function on it This function needs a string variable with the name of the package, it has no return()
}
\examples{

a <- NormData(mean=0, sd =20 ,n=1000, FillUp = FALSE)
newa <- FillUp(a, NA)

}
\seealso{
\code{scale}
}
\author{
Florian Berger <florian_berger@ymail.com>
}
