% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOLdata.R
\name{plotOLdata}
\alias{plotOLdata}
\title{normal plot for simulation data}
\usage{
plotOLdata(data, truevalue, olname, distrest = "fitdistr", logx = TRUE,
  ns = F, diff = F, ylim = NULL, xlim = NULL)
}
\arguments{
\item{data}{list; first sublevel of simulation data as returned by changingn.(...) or changingvar.(...) functions. e.g. simulationdata[['kernel']]}

\item{truevalue}{numeric; a horizontal line with the given number as y-value.}

\item{olname}{string; name of overlap-measure, as returned by OLE function. Must be in the data.}

\item{distrest}{string; determines which data shall be plotted. Determines the first level of the input list and so it can either be 'kernel', 'distrest' or 'normalize'.}

\item{logx}{logical; If TRUE x-axis will be logarithmic.}

\item{ns}{logical; If TRUE the x-axis will be the sample size and not the sublevel of the input data. Can only be TRUE if 'n' is in the data as overlap-measure.}

\item{diff}{logical; If TRUE it will plot the difference to the mean of the chosen datapoints, and not the absolute values.}

\item{ylim}{vector of two numerics; determines the displayed window of the plot on the y axis. will be passed to plot() function. 
@param xlim vector of two numerics; determines the displayed window of the plot on the x axis. will be passed to plot() function.}
}
\value{
nothing.
}
\description{
normal plot for simulation data
}
\details{
takes simulation data, as produced by changingn.randomday as input. Plots the results of the simulation for the Overlapmeasure given by olname and the estimation method (kernel/fitdistr/normalize) given by distrest. The truevalue for it has to be given and will be visible in the plot as horizontal line. this function will dettermine the values for xlim and ylim as the minimum and maximum value of the data on those axes. But they can be given as argument. if logx = TRUE the x-axis will be logarithmic. Will plot the means and give the Standrad deviation as error bar. If diff = TRUE it will plot hte difference to the mean of the pdf-based Overlap-measures.
}
\seealso{
\code{scale}
}
\author{
Florian Berger <florian_berger@ymail.com>
}
