% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sameXs.R
\name{sameXs}
\alias{sameXs}
\title{Creates two datasets with same x-values}
\usage{
sameXs(data1, data2)
}
\arguments{
\item{data1}{data.frame; dataset, which shall have the same x-values (Dates) as data2. It must have two columns, one named 'x' and one named 'y' both containing only numeric values.}

\item{data2}{data.frame; dataset, which shall have the same x-values (Dates) as data1. It must have two columns, one named 'x' and one named 'y' both containing only numeric values.}
}
\value{
a list of data.frames. $data1 is the same as data1 (input) but has in addition the x-values (Dates) of data2 (input), and vice-versa.
}
\description{
Creates two datasets with same x-values
}
\details{
Takes two datasets, it will return a list, of two data-frames, which now have the same x-values. The function will linearly interpolate the x-values for all 'xs' (from data1 and data2) for both data. the input data need to have two columns, one called x and one called y. those should include only numeric values.
}
\examples{

a     <- NormData(mean=0 , sd=5, n = 100)
b     <- NormData(mean=50, sd=5, n = 100)
a1    <- kernelest(a)$data
b1    <- kernelest(b)$data
dats  <- sameXs(a1,b1)
a1new <- dats[[1]]
b1new <- dats[[2]]

}
\seealso{
\code{scale}
}
\author{
Florian Berger <florian_berger@ymail.com>
}
