% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform.R
\name{uniform}
\alias{uniform}
\title{Fits uniform distribution to abundance data}
\usage{
uniform(data, CheckData = TRUE)
}
\arguments{
\item{data}{data.frame;  abundance data, with one column of date objects and one with numeric values.}

\item{CheckData}{logical; if TRUE CheckData function will check format of input data.frame}
}
\value{
returns a list with the levels: $func = a function object, that is created by this function using the dunif() function. $estimationmethod = 'uniform', $startpoint = minimum of dates column in input dataframe, $endpoint = maximum of dates column in input dataframe.
}
\description{
Fits uniform distribution to abundance data (data.frame)
}
\details{
Fits uniform distribution to abundance data, by using the dunif() function and giving it the minimum and maximum dates of the input dataframe.
}
\examples{
a     <- c(1:10)
dates <- as.Date(a, origin = '2017-01-01')
count <- c(1,1,3,4,6,9,5,4,2,1)
dat   <- data.frame(dates, count)
uniform(dat)

}
\seealso{
\code{scale}
}
\author{
Florian Berger <florian_berger@ymail.com>
}
